requirements.txt
asgiref==3.11.0
certifi==2026.1.4
cloudinary==1.44.1
dj-database-url==3.1.0
Django==5.2.10
django-cors-headers==4.9.0
django-jazzmin==3.0.1
djangorestframework==3.16.1
djangorestframework_simplejwt==5.5.1
gunicorn==25.0.3
packaging==26.0
pillow==11.0.0
psycopg2-binary==2.9.11
PyJWT==2.11.0
python-dotenv==1.2.1
six==1.17.0
sqlparse==0.5.5
tzdata==2025.3
urllib3==2.6.3
whitenoise==6.11.0
